Add router w basic routes
