+++ package.json
@@ -13,12 +13,14 @@
     "@types/node": "12.19.9",
     "@types/react": "16.14.2",
     "@types/react-dom": "16.9.10",
+    "@types/react-router-dom": "5.1.6",
     "boardgame.io": "0.42.2",
     "esm": "3.2.25",
     "koa-static": "5.0.0",
     "ramda": "0.27.1",
     "react": "17.0.1",
     "react-dom": "17.0.1",
+    "react-router-dom": "5.2.0",
     "react-scripts": "4.0.1",
     "three": "0.123.0",
     "typescript": "4.1.3",
